FirstStack/requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_iam
aws_cdk.aws_autoscaling
aws_cdk.aws_elasticloadbalancingv2
-e .
